backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pdfplumber
pypdf
python-docx